Reduce ration nutrition by about 37%.
To make up for the increase in permafood from merging the food types.
Perhaps it would be better (fewer keypresses) to instead keep the total
nutrition and decrease the spawn rate, but that (1) is harder to do, (2)
increases variance, and (3) makes it harder to tune food vaults, since
the quantum of permafood would be coarser.